Add tie-break cost to favor fewer redistributions (#89)
Imagine we start node_i from S(0)S(0) and we want to reach node_{i+2} at RR, and that node_{i+1} is an op with zero cost (like alias).
In this case, all of the following chains yield the same cost:
- S(0)S(0) -> S(0)R -> RR
- S(0)S(0) -> RS(0) -> RR
- S(0)S(0) -> RR -> RR
- S(0)S(0) -> S(0)S(0) -> RR
All of those cases are in principle equivalent. But if we want to optimize S(0)S(0) -> RR to dispatch into a single collective
instead of two then we need to favor the last two cases where redistribution happens
in a single go. To do this, we add a tie-break cost that is 1 if a redistribution
happens prior to getting to this configuration, and 0 otherwise. This way,
we will favor having fewer redistributions happening in the graph.

Co-authored-by: Will Constable <whc@meta.com>