Skip placement order optimization for parameters that don't have grad… (#175)
* Skip placement order optimization for parameters that don't have gradients

Previously, we would just error because we would try to access all_input_nodes of grad, which is None. This fixes it by skipping those optimizations for those parameters. In the future, we should make this more robust.

* [WIP] Add tests by Claude

Needed a few tweaks. Still needs cleanup

* Cleanup

* Fix bug in to_fake conversion

If we had parameters that didn't require gradient, this would have been lost during conversion

* Bugfix following codereview